+++ internal/provider/testdata/bigquery_datamart_definition/truncate_with_before_load.tf
@@ -15,8 +15,8 @@ resource "trocco_connection" "my_conn" {
   JSON
 }
 
-resource "trocco_bigquery_datamart_definition" "test_bigquery_datamart" {
-  name                     = "test_bigquery_datamart"
+resource "trocco_bigquery_datamart_definition" "test_truncate_with_before_load" {
+  name                     = "test_truncate_with_before_load"
   is_runnable_concurrently = false
   bigquery_connection_id   = trocco_connection.my_conn.id
   query                    = <<SQL
@@ -29,5 +29,5 @@ resource "trocco_bigquery_datamart_definition" "test_bigquery_datamart" {
   SQL
   destination_dataset      = "dist_datasets"
   destination_table        = "dist_tables"
-  write_disposition        = "append"
+  write_disposition        = "truncate"
 }